services/payment/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.9
alembic==1.10.0
python-jose==3.3.0
passlib>=1.7.4
bcrypt<4.0
pydantic==1.10.7
requests==2.31.0
pika==1.3.1
